fix(admin): Improve table layouts and move Pagination component to separate file
