+++ src/admin/components/index.ts
@@ -1,6 +1,6 @@
 export * from './AdminHeader';
 export * from './AdminTabPanel';
-export * from './Table';
+export * from './table/Table';
 export * from './Notices';
 export * from './controls';
 export * from './contexts';
