+++ src/admin/components/transactions/TransactionsTable.tsx
@@ -1,6 +1,6 @@
 import { Button, Dashicon, Flex, VisuallyHidden } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import { Table } from '../Table';
+import { Table } from '../table/Table';
 import React from 'react';
 import { dateI18n } from '@wordpress/date';
 import { useAdminContext, usePostsContext } from '../contexts';
@@ -31,14 +31,13 @@ export const TransactionsTable = (): React.ReactNode => {
 			key: 'description',
 			title: __('Description', 'kudos-donations'),
 			orderby: 'title',
-			width: '20%',
+			width: '25%',
 			valueCallback: (post: Transaction): React.ReactNode =>
 				post.title.raw,
 		},
 		{
 			key: 'donor',
 			title: __('Donor', 'kudos-donations'),
-			width: '15%',
 			valueCallback: (post: Transaction): React.ReactNode =>
 				post.donor?.meta.email ??
 				post.donor?.meta?.name ??
@@ -49,7 +48,6 @@ export const TransactionsTable = (): React.ReactNode => {
 			key: 'value',
 			orderby: 'meta_value_num',
 			title: __('Amount', 'kudos-donations'),
-			width: '5%',
 			valueCallback: (post: Transaction): React.ReactNode => {
 				const value = post.meta?.value;
 				const currency = post.meta?.currency;
@@ -95,7 +93,6 @@ export const TransactionsTable = (): React.ReactNode => {
 		{
 			key: 'campaign',
 			title: __('Campaign', 'kudos-donations'),
-			width: '20%',
 			valueCallback: (post: Transaction): React.ReactNode =>
 				post.campaign?.title.raw ?? '',
 		},
@@ -103,7 +100,6 @@ export const TransactionsTable = (): React.ReactNode => {
 			key: 'status',
 			orderby: 'meta_value',
 			title: __('Status', 'kudos-donations'),
-			width: '5%',
 			valueCallback: (post: Transaction): React.ReactNode => {
 				const status = post.meta.status;
 
@@ -164,16 +160,9 @@ export const TransactionsTable = (): React.ReactNode => {
 			},
 		},
 		{
-			key: 'message',
-			title: __('Message', 'kudos-donations'),
-			width: '20%',
-			valueCallback: (post: Transaction): string => post.meta.message,
-		},
-		{
 			key: 'date',
 			title: __('Created', 'kudos-donations'),
 			orderby: 'date',
-			width: '10%',
 			valueCallback: (post: Transaction): React.ReactNode => (
 				<i>{dateI18n('d-m-Y', post.date, null)}</i>
 			),
@@ -185,7 +174,6 @@ export const TransactionsTable = (): React.ReactNode => {
 					{__('Actions', 'kudos-donations')}
 				</VisuallyHidden>
 			),
-			width: '5%',
 			valueCallback: (post: Transaction): React.ReactNode => {
 				const status = post.meta.status;
 				const url = post.invoice_url;
