+++ src/admin/kudos-admin.css
@@ -143,7 +143,7 @@
 	}
 
 	.admin-wrap {
-		max-width: 1080px;
+		max-width: 900px;
 		margin: auto;
 	}
 
@@ -188,6 +188,13 @@
 			opacity: 0.5;
 			pointer-events: none;
 			transition: opacity 0.2s ease-in-out;
+		}
+
+		.table-cell {
+			vertical-align: middle;
+			white-space: normal;
+			word-break: break-word;
+			overflow-wrap: anywhere;
 		}
 
 		th,td {
